DeepTalk/generate_cert.py
#!/usr/bin/env python3
"""Generate self-signed SSL certificate for local development."""

import os
from cryptography import x509
from cryptography.x509.oid import NameOID
from cryptography.hazmat.primitives import hashes
from cryptography.hazmat.primitives.asymmetric import rsa
from cryptography.hazmat.primitives import serialization
import datetime
import ipaddress

def generate_self_signed_cert():
    # Generate private key
    private_key = rsa.generate_private_key(
        public_exponent=65537,
        key_size=2048,
    )

    # Get local IP
    import socket
    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
    s.connect(('8.8.8.8', 80))
    local_ip = s.getsockname()[0]
    s.close()

    # Create certificate
    subject = issuer = x509.Name([
        x509.NameAttribute(NameOID.COUNTRY_NAME, u"DE"),
        x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, u"Local"),
        x509.NameAttribute(NameOID.LOCALITY_NAME, u"DeepTalk"),
        x509.NameAttribute(NameOID.ORGANIZATION_NAME, u"DeepTalk Dev"),
        x509.NameAttribute(NameOID.COMMON_NAME, u"localhost"),
    ])

    cert = x509.CertificateBuilder().subject_name(
        subject
    ).issuer_name(
        issuer
    ).public_key(
        private_key.public_key()
    ).serial_number(
        x509.random_serial_number()
    ).not_valid_before(
        datetime.datetime.utcnow()
    ).not_valid_after(
        datetime.datetime.utcnow() + datetime.timedelta(days=365)
    ).add_extension(
        x509.SubjectAlternativeName([
            x509.DNSName(u"localhost"),
            x509.DNSName(u"*.localhost"),
            x509.IPAddress(ipaddress.IPv4Address(u"127.0.0.1")),
            x509.IPAddress(ipaddress.IPv4Address(local_ip)),
        ]),
        critical=False,
    ).sign(private_key, hashes.SHA256())

    # Write certificate and key files
    cert_path = "deeptalk.crt"
    key_path = "deeptalk.key"

    with open(cert_path, "wb") as f:
        f.write(cert.public_bytes(serialization.Encoding.PEM))

    with open(key_path, "wb") as f:
        f.write(private_key.private_bytes(
            encoding=serialization.Encoding.PEM,
            format=serialization.PrivateFormat.PKCS8,
            encryption_algorithm=serialization.NoEncryption()
        ))

    print(f"✅ Generated SSL certificate: {cert_path}")
    print(f"✅ Generated SSL private key: {key_path}")
    print(f"✅ Certificate valid for: localhost, 127.0.0.1, {local_ip}")
    
    return cert_path, key_path

if __name__ == "__main__":
    try:
        generate_self_signed_cert()
        print("\n🔒 SSL certificates created successfully!")
        print("You can now run Flask with HTTPS support.")
    except ImportError:
        print("❌ Missing cryptography package. Installing it...")
        print("Run: pip install cryptography")
    except Exception as e:
        print(f"❌ Error generating certificates: {e}")