DeepTalk/requirements.txt
flask==3.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flask-sock==0.7.0
itsdangerous==2.2.0
